refactor: :recycle: Refactor test structure and alphaburned event examples
